Use ephemeral OTA Provider instances
Instead of trying to reuse the same OTA Provider instance for multiple
OTA requests, create a new instance for each request. This is easier
to implement and also allows parallel updates.

Because the OTA Provider is now ephemeral, we need to commission it on
every update. But this is quick and reliable, so not a big deal.

To support multiple updates at once, we need to make sure the OTA
Providers use a distinct Matter port (hence passing 0) and distinct
node ids. The current implementation simply uses the target node id
plus a fixed offset. Since a single node can only run one update at a
time, this is sufficient.

Furthermore, some updates seem to have a difference in reported
versionNumberString value in the DCL vs. what is actually in the OTA
metadata. Specifically Eve updates from the Testnet DCL are such
updates (e.g. 3.2.0 vs 3.2.6705). When using the OTA Provider with the
--otaImageList option, this discrepancy is an issue and causes OTA
Provider to abort.

Using the single OTA update per OTA Provider instance allows us to use
--filepath, which doesn't check the versionNumberString.